apps/api-gateway/requirements.txt
# ── Web Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# ── Database ──
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# ── Validation & Settings ──
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# ── Auth ──
PyJWT==2.10.1
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
aiosmtplib==5.1.0
dnspython==2.8.0
email-validator==2.3.0
bcrypt-4.2.1

# ── ML: Core ──
scikit-learn==1.6.0
joblib==1.4.2
numpy>=2.1.0

# ── ML: Audio ──
librosa==0.10.2.post1
soundfile==0.12.1
tensorflow>=2.16.0

# ── ML: Video ──
opencv-python-headless==4.10.0.84
mediapipe>=0.10.30

# ── ML: QR ──
pyzbar==0.1.9
Pillow==11.1.0

# ── Utilities ──
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
tqdm>=4.66.0
feedparser>=6.0.0

# ── Testing ──
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# ── ML / Deepfake Detection ──
torch>=2.0.0
torchvision>=0.15.0
librosa>=0.10.0
soundfile>=0.12.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
joblib>=1.3.0